Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int m;
+    float m;
 
     scanf("%f",&m);
     printf("\nYou entered: %.2f",m);
